git-svn-id: http://192.168.20.11/sequoiadb/trunk@10760 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/rtn/rtnMsg.cpp
@@ -11,7 +11,7 @@ namespace engine
    {
       INT32 rc = SDB_OK ;
 
-      if ( pMsg->header.messageLength < sizeof( MsgHeader ) )
+      if ( (UINT32)pMsg->header.messageLength < sizeof( MsgHeader ) )
       {
          PD_LOG( PDERROR, "Recieve invalid msg[length: %d]",
                  pMsg->header.messageLength ) ;
